Add extreme/overload bulkheads to test level

+++ AutogenRundown/src/DataBlocks/Objectives/EventBuilder.cs
@@ -25,7 +25,7 @@
                 {
                     Type = WardenObjectiveEventType.PlaySound,
                     Trigger = WardenObjectiveEventTrigger.OnStart,
-                    SoundId = (uint)Sound.LightsOff,
+                    SoundId = Sound.LightsOff,
                     WardenIntel = message,
                     Delay = delay
                 });
